+++ lib/ts/recipe/passwordless/components/themes/signInUpEPCombo/emailOrPhoneForm.tsx
@@ -109,10 +109,13 @@ export const EPComboEmailOrPhoneForm = withOverride(
                         <Label value={"PWLESS_COMBO_PASSWORD_LABEL"} data-supertokens="passwordInputLabel" />
                         <a
                             onClick={() =>
-                                EmailPassword.getInstanceOrThrow().redirect({
-                                    action: "RESET_PASSWORD",
-                                    tenantIdFromQueryParams: getTenantIdFromQueryParams(),
-                                })
+                                EmailPassword.getInstanceOrThrow().redirect(
+                                    {
+                                        action: "RESET_PASSWORD",
+                                        tenantIdFromQueryParams: getTenantIdFromQueryParams(),
+                                    },
+                                    props.navigate
+                                )
                             }
                             data-supertokens="link linkButton formLabelLinkBtn forgotPasswordLink">
                             {t("PWLESS_COMBO_FORGOT_PW_LINK")}